Flatten directory structure. Move test applications and DAQModules to test subdirectory.
