+++ apps/CMakeLists.txt
@@ -1,14 +1,6 @@
 include_directories(${CMAKE_SOURCE_DIR}/app-framework/include)
 include_directories(${CMAKE_SOURCE_DIR}/app-framework-base/include)
 
-add_executable(echo_test_app echo_test_app.cc)
-target_link_libraries(echo_test_app app-framework app-framework_DebugLoggingDAQModule_duneDAQModule pthread ers  ${CETLIB} ${CETLIB_EXCEPT} ${Boost_PROGRAM_OPTIONS_LIBRARY})
+add_executable(daq_application daq_application.cc)
+target_link_libraries(daq_application app-framework  pthread ers ${CETLIB} ${CETLIB_EXCEPT} ${Boost_PROGRAM_OPTIONS_LIBRARY})
 
-add_executable(producer_consumer_test_app producer_consumer_test_app.cc)
-target_link_libraries(producer_consumer_test_app app-framework app-framework_FakeDataProducerDAQModule_duneDAQModule ers app-framework_FakeDataConsumerDAQModule_duneDAQModule pthread  ${CETLIB} ${CETLIB_EXCEPT} ${Boost_PROGRAM_OPTIONS_LIBRARY})
-
-add_executable(dynamic_loading_test_app dynamic_loading_test_app.cc)
-target_link_libraries(dynamic_loading_test_app app-framework  pthread ers ${CETLIB} ${CETLIB_EXCEPT} ${Boost_PROGRAM_OPTIONS_LIBRARY})
-
-add_executable(daq_sink_source daq_sink_source.cc)
-target_link_libraries(daq_sink_source app-framework pthread ers)
