+++ include/app-framework/StdDeQueue.hh
