+++ include/app-framework/detail/StdDeQueue.icc
