+++ src/CMakeLists.txt
@@ -5,7 +5,6 @@ list(FILTER source_files EXCLUDE REGEX "QueryResponseCommandFacility.cc")
 add_library(app-framework SHARED ${source_files})
 target_link_libraries(app-framework ${CETLIB} ${CETLIB_EXCEPT})
 
-add_subdirectory(DAQModules)
 
 add_library(app-framework_QueryResponseCommandFacility_duneCommandFacility QueryResponseCommandFacility.cc)
