--- src/DAQModules/CMakeLists.txt
@@ -1,7 +1,0 @@
-
-#file(GLOB source_files *.cc)
-#add_library(app-framework_DAQModules SHARED ${source_files} )
-
-add_library(app-framework_DebugLoggingDAQModule_duneDAQModule DebugLoggingDAQModule.cc)
-add_library(app-framework_FakeDataConsumerDAQModule_duneDAQModule FakeDataConsumerDAQModule.cc)
-add_library(app-framework_FakeDataProducerDAQModule_duneDAQModule FakeDataProducerDAQModule.cc)
