+++ src/QueryResponseCommandFacility.hh
@@ -14,7 +14,7 @@
 
 #include "app-framework/DAQProcess.hh"
 
-#include "app-framework/Services/CommandFacility.hh"
+#include "app-framework/CommandFacility.hh"
 
 namespace appframework {
 /**
