+++ unittest/DAQModuleThreadHelper_test.cc
@@ -7,7 +7,7 @@
  * received with this code.
  */
 
-#include "app-framework/DAQModules/DAQModuleThreadHelper.hh"
+#include "app-framework/DAQModuleThreadHelper.hh"
 
 #define BOOST_TEST_MODULE DAQModuleThreadHelper_test // NOLINT
 #include <boost/test/unit_test.hpp>
